RSPY-798: Moved code
